% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the funresMech Shiny App}
\usage{
run_app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}
(e.g., \code{port}, \code{host}, \code{launch.browser}).}
}
\value{
This function launches a Shiny app and does not return a value.
}
\description{
Launches the interactive application for mechanistic functional response
analysis, based on the Okuyama model.
}
\examples{
if (interactive()) {
  run_app()
}

}
